Break the client up into a module
