+++ pythosf/__init__.py
@@ -1,1 +1,3 @@
-__all__ = ['client', 'exceptions']
+from pythosf.utils import bearer_token_auth
+
+__all__ = ['client', 'exceptions', 'bearer_token_auth', ]
