+++ pythosf/client/provider.py
@@ -1,0 +1,6 @@
+from .folder import Folder
+
+
+class Provider(Folder):
+    pass
+
